Move shardscanner workflow tests to the shardscanner package
This ensures they get included in the code coverage metrics.
